Increase maintainability by splitting up headers. Change CAN numbering to sequential address space. (0,1,2, not 0,1,4) Lay foundation for setting canbus speed.
