+++ board/Makefile.legacy
@@ -1,6 +1,6 @@
 # :set noet
 PROJ_NAME = comma
-CFLAGS  = -g -O0 -Wall
+CFLAGS  = -g -O0 -Wall -std=gnu90
 
 CFLAGS += -mlittle-endian -mthumb -mcpu=cortex-m3
 CFLAGS += -msoft-float -DSTM32F2 -DSTM32F205xx
