+++ board/main.c
@@ -1,30 +1,18 @@
 #include "config.h"
 #include "early.h"
-
-#define NULL ((void*)0)
-
-
-// assign CAN numbering
-// old:         CAN1 = 1   CAN2 = 0
-// panda:       CAN1 = 0   CAN2 = 1   CAN3 = 4
-#ifdef PANDA
-  int can_numbering[] = {0,1,4};
-#else
-  int can_numbering[] = {1,0,-1};
-#endif
-
-// can forwards FROM -> TO
-#define CAN_MAX 3
-int can_forwarding[] = {-1,-1,-1};
-
-// *** end config ***
-
 #include "obj/gitversion.h"
+#include "uart.h"
+#include "can.h"
+#include "libc.h"
+#include "gpio.h"
+#include "adc.h"
+#include "timer.h"
+#include "usb.h"
+#include "spi.h"
+#include "safety.h"
 
 // debug safety check: is controls allowed?
-int controls_allowed = 0;
 int started = 0;
-int can_live = 0, pending_can_live = 0;
 
 // optional features
 int gas_interceptor_detected = 0;
@@ -34,106 +22,11 @@ int started_signal_detected = 0;
 // TODO: check for UART high
 int did_usb_enumerate = 0;
 
-// Declare puts to supress warning
-int puts ( const char * str );
-
-// ********************* instantiate queues *********************
-
-typedef struct {
-  uint32_t w_ptr;
-  uint32_t r_ptr;
-  uint32_t fifo_size;
-  CAN_FIFOMailBox_TypeDef *elems;
-} can_ring;
-
-#define can_buffer(x, size) \
-  CAN_FIFOMailBox_TypeDef elems_##x[size]; \
-  can_ring can_##x = { .w_ptr = 0, .r_ptr = 0, .fifo_size = size, .elems = (CAN_FIFOMailBox_TypeDef *)&elems_##x };
-
-can_buffer(rx_q, 0x1000)
-can_buffer(tx1_q, 0x100)
-can_buffer(tx2_q, 0x100)
-can_buffer(tx3_q, 0x100)
-
-// ********************* interrupt safe queue *********************
-
-int pop(can_ring *q, CAN_FIFOMailBox_TypeDef *elem) {
-  if (q->w_ptr != q->r_ptr) {
-    *elem = q->elems[q->r_ptr];
-    if ((q->r_ptr + 1) == q->fifo_size) q->r_ptr = 0;
-    else q->r_ptr += 1;
-    return 1;
-  }
-  return 0;
-}
-
-int push(can_ring *q, CAN_FIFOMailBox_TypeDef *elem) {
-  uint32_t next_w_ptr;
-  if ((q->w_ptr + 1) == q->fifo_size) next_w_ptr = 0;
-  else next_w_ptr = q->w_ptr + 1;
-  if (next_w_ptr != q->r_ptr) {
-    q->elems[q->w_ptr] = *elem;
-    q->w_ptr = next_w_ptr;
-    return 1;
-  }
-  puts("push failed!\n");
-  return 0;
-}
-
-// ***************************** serial port queues *****************************
-
-#define FIFO_SIZE 0x100
-
-typedef struct uart_ring {
-  uint8_t w_ptr_tx;
-  uint8_t r_ptr_tx;
-  uint8_t elems_tx[FIFO_SIZE];
-  uint8_t w_ptr_rx;
-  uint8_t r_ptr_rx;
-  uint8_t elems_rx[FIFO_SIZE];
-  USART_TypeDef *uart;
-  void (*callback)(struct uart_ring*);
-} uart_ring;
-
-int getc(uart_ring *q, char *elem);
-int putc(uart_ring *q, char elem);
-
-// esp = USART1
-uart_ring esp_ring = { .w_ptr_tx = 0, .r_ptr_tx = 0,
-                       .w_ptr_rx = 0, .r_ptr_rx = 0,
-                       .uart = USART1 };
-
-// lin1, K-LINE = UART5
-// lin2, L-LINE = USART3
-uart_ring lin1_ring = { .w_ptr_tx = 0, .r_ptr_tx = 0,
-                        .w_ptr_rx = 0, .r_ptr_rx = 0,
-                        .uart = UART5 };
-uart_ring lin2_ring = { .w_ptr_tx = 0, .r_ptr_tx = 0,
-                        .w_ptr_rx = 0, .r_ptr_rx = 0,
-                        .uart = USART3 };
-
-// debug = USART2
-void debug_ring_callback(uart_ring *ring);
-uart_ring debug_ring = { .w_ptr_tx = 0, .r_ptr_tx = 0,
-                         .w_ptr_rx = 0, .r_ptr_rx = 0,
-                         .uart = USART2,
-                         .callback = debug_ring_callback};
-
-
-uart_ring *get_ring_by_number(int a) {
-  switch(a) {
-    case 0:
-      return &debug_ring;
-    case 1:
-      return &esp_ring;
-    case 2:
-      return &lin1_ring;
-    case 3:
-      return &lin2_ring;
-    default:
-      return NULL;
-  }
-}
+#ifdef PANDA_SAFETY
+#include "panda_safety.h"
+#else
+#include "honda_safety.h"
+#endif
 
 void accord_framing_callback(uart_ring *q) {
   uint8_t r_ptr_rx_tmp = q->r_ptr_rx;
@@ -196,264 +89,6 @@ void accord_framing_callback(uart_ring *q) {
   }
 }
 
-void debug_ring_callback(uart_ring *ring) {
-  char rcv;
-  while (getc(ring, &rcv)) {
-    putc(ring, rcv);
-
-    // jump to DFU flash
-    if (rcv == 'z') {
-      enter_bootloader_mode = ENTER_BOOTLOADER_MAGIC;
-      NVIC_SystemReset();
-    }
-    if (rcv == 'x') {
-      // normal reset
-      NVIC_SystemReset();
-    }
-  }
-}
-
-// ***************************** serial port *****************************
-
-void uart_ring_process(uart_ring *q) {
-  // TODO: check if external serial is connected
-  int sr = q->uart->SR;
-
-  if (q->w_ptr_tx != q->r_ptr_tx) {
-    if (sr & USART_SR_TXE) {
-      q->uart->DR = q->elems_tx[q->r_ptr_tx];
-      q->r_ptr_tx += 1;
-    } else {
-      // push on interrupt later
-      q->uart->CR1 |= USART_CR1_TXEIE;
-    }
-  } else {
-    // nothing to send
-    q->uart->CR1 &= ~USART_CR1_TXEIE;
-  }
-
-  if (sr & USART_SR_RXNE) {
-    uint8_t c = q->uart->DR;  // TODO: can drop packets
-    uint8_t next_w_ptr = q->w_ptr_rx + 1;
-    if (next_w_ptr != q->r_ptr_rx) {
-      q->elems_rx[q->w_ptr_rx] = c;
-      q->w_ptr_rx = next_w_ptr;
-      if (q->callback) q->callback(q);
-    }
-  }
-}
-
-// interrupt boilerplate
-
-void USART1_IRQHandler(void) {
-  NVIC_DisableIRQ(USART1_IRQn);
-  uart_ring_process(&esp_ring);
-  NVIC_EnableIRQ(USART1_IRQn);
-}
-
-void USART2_IRQHandler(void) {
-  NVIC_DisableIRQ(USART2_IRQn);
-  uart_ring_process(&debug_ring);
-  NVIC_EnableIRQ(USART2_IRQn);
-}
-
-void USART3_IRQHandler(void) {
-  NVIC_DisableIRQ(USART3_IRQn);
-  uart_ring_process(&lin2_ring);
-  NVIC_EnableIRQ(USART3_IRQn);
-}
-
-void UART5_IRQHandler(void) {
-  NVIC_DisableIRQ(UART5_IRQn);
-  uart_ring_process(&lin1_ring);
-  NVIC_EnableIRQ(UART5_IRQn);
-}
-
-int getc(uart_ring *q, char *elem) {
-  if (q->w_ptr_rx != q->r_ptr_rx) {
-    *elem = q->elems_rx[q->r_ptr_rx];
-    q->r_ptr_rx += 1;
-    return 1;
-  }
-  return 0;
-}
-
-int injectc(uart_ring *q, char elem) {
-  uint8_t next_w_ptr = q->w_ptr_rx + 1;
-  int ret = 0;
-  if (next_w_ptr != q->r_ptr_rx) {
-    q->elems_rx[q->w_ptr_rx] = elem;
-    q->w_ptr_rx = next_w_ptr;
-    ret = 1;
-  }
-  return ret;
-}
-
-int putc(uart_ring *q, char elem) {
-  uint8_t next_w_ptr = q->w_ptr_tx + 1;
-  int ret = 0;
-  if (next_w_ptr != q->r_ptr_tx) {
-    q->elems_tx[q->w_ptr_tx] = elem;
-    q->w_ptr_tx = next_w_ptr;
-    ret = 1;
-  }
-  uart_ring_process(q);
-  return ret;
-}
-
-// ********************* includes *********************
-
-#include "libc.h"
-#include "gpio.h"
-#include "uart.h"
-#include "adc.h"
-#include "timer.h"
-#include "usb.h"
-#include "can.h"
-#include "spi.h"
-
-void safety_rx_hook(CAN_FIFOMailBox_TypeDef *to_push);
-int safety_tx_hook(CAN_FIFOMailBox_TypeDef *to_send, int hardwired);
-int safety_tx_lin_hook(int lin_num, uint8_t *data, int len, int hardwired);
-
-#ifdef PANDA_SAFETY
-#include "panda_safety.h"
-#else
-#include "honda_safety.h"
-#endif
-
-// ***************************** CAN *****************************
-
-void process_can(CAN_TypeDef *CAN, can_ring *can_q, int can_number) {
-  #ifdef DEBUG
-    puts("process CAN TX\n");
-  #endif
-
-  // add successfully transmitted message to my fifo
-  if ((CAN->TSR & CAN_TSR_TXOK0) == CAN_TSR_TXOK0) {
-    CAN_FIFOMailBox_TypeDef to_push;
-    to_push.RIR = CAN->sTxMailBox[0].TIR;
-    to_push.RDTR = (CAN->sTxMailBox[0].TDTR & 0xFFFF000F) | ((can_number+2) << 4);
-    to_push.RDLR = CAN->sTxMailBox[0].TDLR;
-    to_push.RDHR = CAN->sTxMailBox[0].TDHR;
-    push(&can_rx_q, &to_push);
-  }
-
-  // check for empty mailbox
-  CAN_FIFOMailBox_TypeDef to_send;
-  if ((CAN->TSR & CAN_TSR_TME0) == CAN_TSR_TME0) {
-    if (pop(can_q, &to_send)) {
-      // only send if we have received a packet
-      CAN->sTxMailBox[0].TDLR = to_send.RDLR;
-      CAN->sTxMailBox[0].TDHR = to_send.RDHR;
-      CAN->sTxMailBox[0].TDTR = to_send.RDTR;
-      CAN->sTxMailBox[0].TIR = to_send.RIR;
-    }
-  }
-
-  // clear interrupt
-  CAN->TSR |= CAN_TSR_RQCP0;
-}
-
-// send more, possible for these to not trigger?
-
-
-void CAN1_TX_IRQHandler() {
-  process_can(CAN1, &can_tx1_q, can_numbering[0]);
-}
-
-void CAN2_TX_IRQHandler() {
-  process_can(CAN2, &can_tx2_q, can_numbering[1]);
-}
-
-#ifdef CAN3
-void CAN3_TX_IRQHandler() {
-  process_can(CAN3, &can_tx3_q, can_numbering[2]);
-}
-#endif
-
-void send_can(CAN_FIFOMailBox_TypeDef *to_push, int flags);
-
-// CAN receive handlers
-// blink blue when we are receiving CAN messages
-void can_rx(CAN_TypeDef *CAN, int can_index) {
-  int can_number = can_numbering[can_index];
-  while (CAN->RF0R & CAN_RF0R_FMP0) {
-    // can is live
-    pending_can_live = 1;
-
-    // add to my fifo
-    CAN_FIFOMailBox_TypeDef to_push;
-    to_push.RIR = CAN->sFIFOMailBox[0].RIR;
-    to_push.RDTR = CAN->sFIFOMailBox[0].RDTR;
-    to_push.RDLR = CAN->sFIFOMailBox[0].RDLR;
-    to_push.RDHR = CAN->sFIFOMailBox[0].RDHR;
-
-    // forwarding (panda only)
-    #ifdef PANDA
-      if (can_forwarding[can_index] != -1 && can_numbering[can_forwarding[can_index]] != -1) {
-        CAN_FIFOMailBox_TypeDef to_send;
-        to_send.RIR = to_push.RIR | 1; // TXRQ
-        to_send.RDTR = to_push.RDTR;
-        to_send.RDLR = to_push.RDLR;
-        to_send.RDHR = to_push.RDHR;
-        send_can(&to_send, can_numbering[can_forwarding[can_index]]);
-      }
-    #endif
-
-    // modify RDTR for our API
-    to_push.RDTR = (to_push.RDTR & 0xFFFF000F) | (can_number << 4);
-
-    safety_rx_hook(&to_push);
-
-    #ifdef PANDA
-      set_led(LED_GREEN, 1);
-    #endif
-    push(&can_rx_q, &to_push);
-
-    // next
-    CAN->RF0R |= CAN_RF0R_RFOM0;
-  }
-}
-
-void CAN1_RX0_IRQHandler() {
-  //puts("CANRX1");
-  //delay(10000);
-  can_rx(CAN1, 0);
-}
-
-void CAN2_RX0_IRQHandler() {
-  //puts("CANRX0");
-  //delay(10000);
-  can_rx(CAN2, 1);
-}
-
-#ifdef CAN3
-void CAN3_RX0_IRQHandler() {
-  //puts("CANRX0");
-  //delay(10000);
-  can_rx(CAN3, 2);
-}
-#endif
-
-void CAN1_SCE_IRQHandler() {
-  //puts("CAN1_SCE\n");
-  can_sce(CAN1);
-}
-
-void CAN2_SCE_IRQHandler() {
-  //puts("CAN2_SCE\n");
-  can_sce(CAN2);
-}
-
-#ifdef CAN3
-void CAN3_SCE_IRQHandler() {
-  //puts("CAN3_SCE\n");
-  can_sce(CAN3);
-}
-#endif
-
-
 // ***************************** USB port *****************************
 
 int get_health_pkt(void *dat) {
@@ -511,44 +146,9 @@ void usb_cb_ep2_out(uint8_t *usbdata, int len, int hardwired) {
   }
 }
 
-void send_can(CAN_FIFOMailBox_TypeDef *to_push, int flags) {
-  int i;
-  CAN_TypeDef *CAN;
-  can_ring *can_q;
-  if (flags == can_numbering[0])  {
-    CAN = CAN1;
-    can_q = &can_tx1_q;
-  } else if (flags == can_numbering[1]) {
-    CAN = CAN2;
-    can_q = &can_tx2_q;
-  #ifdef CAN3
-  } else if (flags == can_numbering[2]) {
-    CAN = CAN3;
-    can_q = &can_tx3_q;
-  #endif
-  } else if (flags == 8 || flags == 9) {
-    // fake LIN as CAN
-    uart_ring *lin_ring = (flags == 8) ? &lin1_ring : &lin2_ring;
-    for (i = 0; i < min(8, to_push->RDTR & 0xF); i++) {
-      putc(lin_ring, ((uint8_t*)&to_push->RDLR)[i]);
-    }
-    return;
-  } else {
-    // no crash
-    return;
-  }
-
-  // add CAN packet to send queue
-  // bus number isn't passed through
-  to_push->RDTR &= 0xF;
-  push(can_q, to_push);
-
-  // flags = can_number
-  process_can(CAN, can_q, flags);
-}
-
 // send on CAN
 void usb_cb_ep3_out(uint8_t *usbdata, int len, int hardwired) {
+  puts("usb_cb_ep3_out called\n");
   int dpkt = 0;
   for (dpkt = 0; dpkt < len; dpkt += 0x10) {
     uint32_t *tf = (uint32_t*)(&usbdata[dpkt]);
@@ -641,29 +241,28 @@ int usb_cb_control_msg(USB_Setup_TypeDef *setup, uint8_t *resp, int hardwired) {
       }
       break;
     case 0xdb: // toggle GMLAN
-      if (setup->b.wIndex.w == 3) {
+      if (setup->b.wValue.w == 2) {
+        set_can_mode(1, 0);
+        set_can_mode(2, setup->b.wIndex.w);
+      } else if (setup->b.wValue.w == 1) {
         set_can_mode(2, 0);
-        set_can_mode(3, setup->b.wValue.w);
-      } else {
-        set_can_mode(3, 0);
-        set_can_mode(2, setup->b.wValue.w);
+        set_can_mode(1, setup->b.wIndex.w);
       }
       break;
     case 0xdc: // set controls allowed
       controls_allowed = setup->b.wValue.w == 0x1337;
       // take CAN out of SILM, careful with speed!
-      can_init(CAN1, 0);
-      can_init(CAN2, 0);
+      can_init(0);
+      can_init(1);
       #ifdef CAN3
-        can_init(CAN3, 0);
+      can_init(2);
       #endif
       break;
     case 0xdd: // enable can forwarding
-      if (setup->b.wValue.w != 0 && setup->b.wValue.w <= CAN_MAX) {
-        // 0 sets it to -1
-        if (setup->b.wIndex.w <= CAN_MAX) {
-          can_forwarding[setup->b.wValue.w-1] = setup->b.wIndex.w-1;
-        }
+      if (setup->b.wValue.w < CAN_MAX) { //Set forwarding
+        can_forwarding[setup->b.wValue.w] = setup->b.wIndex.w;
+      }else if(setup->b.wValue.w == 0xFF){ //Clear Forwarding
+        can_forwarding[setup->b.wValue.w] = -1;
       }
       break;
     case 0xe0: // uart read
@@ -751,7 +350,6 @@ int usb_cb_control_msg(USB_Setup_TypeDef *setup, uint8_t *resp, int hardwired) {
       puts("NO HANDLER ");
       puth(setup->b.bRequest);
       puts("\n");
-      break;
   }
   return resp_len;
 }
@@ -848,6 +446,8 @@ void __initialize_hardware_early() {
 }
 
 int main() {
+  int i;
+
   // init devices
   clock_init();
   periph_init();
@@ -867,27 +467,21 @@ int main() {
   uart_init(USART3, 10400);
   USART3->CR2 |= USART_CR2_LINEN;
 
-  /*puts("EXTERNAL");
-  puth(has_external_debug_serial);
-  puts("\n");*/
-
   // enable USB
   usb_init();
 
   // default to silent mode to prevent issues with Ford
 #ifdef PANDA_SAFETY
-  can_init(CAN1, 1);
-  can_init(CAN2, 1);
-  #ifdef CAN3
-    can_init(CAN3, 1);
-  #endif
+  controls_allowed = 0;
 #else
-  can_init(CAN1, 0);
-  can_init(CAN2, 0);
-  #ifdef CAN3
-    can_init(CAN3, 0);
-  #endif
+  controls_allowed = 1;
 #endif
+  puts("Can0 default bitrate ");
+  puth(can_bitrate[0]);
+  puts("\n");
+
+  for(i=0; i < CAN_MAX; i++)
+    can_init(i);
 
   adc_init();
