+++ board/uart.c
@@ -1,0 +1,208 @@
+#include <stdint.h>
+#include "config.h"
+#include "uart.h"
+#include "early.h"
+
+void debug_ring_callback(uart_ring *ring) {
+  char rcv;
+  while (getc(ring, &rcv)) {
+    putc(ring, rcv);
+
+    // jump to DFU flash
+    if (rcv == 'z') {
+      enter_bootloader_mode = ENTER_BOOTLOADER_MAGIC;
+      NVIC_SystemReset();
+    }
+    if (rcv == 'x') {
+      // normal reset
+      NVIC_SystemReset();
+    }
+  }
+}
+
+// ***************************** serial port queues *****************************
+
+uart_ring debug_ring = { .w_ptr_tx = 0, .r_ptr_tx = 0,
+                         .w_ptr_rx = 0, .r_ptr_rx = 0,
+                         .uart = USART2,
+                         .callback = debug_ring_callback};
+
+// esp = USART1
+uart_ring esp_ring = { .w_ptr_tx = 0, .r_ptr_tx = 0,
+                       .w_ptr_rx = 0, .r_ptr_rx = 0,
+                       .uart = USART1 };
+
+// lin1, K-LINE = UART5
+// lin2, L-LINE = USART3
+uart_ring lin1_ring = { .w_ptr_tx = 0, .r_ptr_tx = 0,
+                        .w_ptr_rx = 0, .r_ptr_rx = 0,
+                        .uart = UART5 };
+uart_ring lin2_ring = { .w_ptr_tx = 0, .r_ptr_tx = 0,
+                        .w_ptr_rx = 0, .r_ptr_rx = 0,
+                        .uart = USART3 };
+
+uart_ring *get_ring_by_number(int a) {
+  switch(a) {
+    case 0:
+      return &debug_ring;
+    case 1:
+      return &esp_ring;
+    case 2:
+      return &lin1_ring;
+    case 3:
+      return &lin2_ring;
+    default:
+      return NULL;
+  }
+}
+
+void uart_set_baud(USART_TypeDef *u, int baud) {
+  if (u == USART1) {
+    // USART1 is on APB2
+    u->BRR = __USART_BRR(48000000, baud);
+  } else {
+    u->BRR = __USART_BRR(24000000, baud);
+  }
+}
+
+void uart_init(USART_TypeDef *u, int baud) {
+  // enable uart and tx+rx mode
+  u->CR1 = USART_CR1_UE;
+  uart_set_baud(u, baud);
+
+  u->CR1 |= USART_CR1_TE | USART_CR1_RE;
+  //u->CR2 = USART_CR2_STOP_0 | USART_CR2_STOP_1;
+  //u->CR2 = USART_CR2_STOP_0;
+  // ** UART is ready to work **
+
+  // enable interrupts
+  u->CR1 |= USART_CR1_RXNEIE;
+}
+
+void putch(const char a) {
+  if (has_external_debug_serial) {
+    putc(&debug_ring, a);
+  } else {
+    injectc(&debug_ring, a);
+  }
+}
+
+int puts(const char *a) {
+  for (;*a;a++) {
+    if (*a == '\n') putch('\r');
+    putch(*a);
+  }
+  return 0;
+}
+
+void puth(unsigned int i) {
+  int pos;
+  char c[] = "0123456789abcdef";
+  for (pos = 28; pos != -4; pos -= 4) {
+    putch(c[(i >> pos) & 0xF]);
+  }
+}
+
+void puth2(unsigned int i) {
+  int pos;
+  char c[] = "0123456789abcdef";
+  for (pos = 4; pos != -4; pos -= 4) {
+    putch(c[(i >> pos) & 0xF]);
+  }
+}
+
+void hexdump(void *a, int l) {
+  int i;
+  for (i=0;i<l;i++) {
+    if (i != 0 && (i&0xf) == 0) puts("\n");
+    puth2(((unsigned char*)a)[i]);
+    puts(" ");
+  }
+  puts("\n");
+}
+
+int getc(uart_ring *q, char *elem) {
+  if (q->w_ptr_rx != q->r_ptr_rx) {
+    *elem = q->elems_rx[q->r_ptr_rx];
+    q->r_ptr_rx += 1;
+    return 1;
+  }
+  return 0;
+}
+
+int putc(uart_ring *q, char elem) {
+  uint8_t next_w_ptr = q->w_ptr_tx + 1;
+  int ret = 0;
+  if (next_w_ptr != q->r_ptr_tx) {
+    q->elems_tx[q->w_ptr_tx] = elem;
+    q->w_ptr_tx = next_w_ptr;
+    ret = 1;
+  }
+  uart_ring_process(q);
+  return ret;
+}
+
+int injectc(uart_ring *q, char elem) {
+  uint8_t next_w_ptr = q->w_ptr_rx + 1;
+  int ret = 0;
+  if (next_w_ptr != q->r_ptr_rx) {
+    q->elems_rx[q->w_ptr_rx] = elem;
+    q->w_ptr_rx = next_w_ptr;
+    ret = 1;
+  }
+  return ret;
+}
+
+void uart_ring_process(uart_ring *q) {
+  // TODO: check if external serial is connected
+  int sr = q->uart->SR;
+
+  if (q->w_ptr_tx != q->r_ptr_tx) {
+    if (sr & USART_SR_TXE) {
+      q->uart->DR = q->elems_tx[q->r_ptr_tx];
+      q->r_ptr_tx += 1;
+    } else {
+      // push on interrupt later
+      q->uart->CR1 |= USART_CR1_TXEIE;
+    }
+  } else {
+    // nothing to send
+    q->uart->CR1 &= ~USART_CR1_TXEIE;
+  }
+
+  if (sr & USART_SR_RXNE) {
+    uint8_t c = q->uart->DR;  // TODO: can drop packets
+    uint8_t next_w_ptr = q->w_ptr_rx + 1;
+    if (next_w_ptr != q->r_ptr_rx) {
+      q->elems_rx[q->w_ptr_rx] = c;
+      q->w_ptr_rx = next_w_ptr;
+      if (q->callback) q->callback(q);
+    }
+  }
+}
+
+// interrupt boilerplate
+
+void USART1_IRQHandler(void) {
+  NVIC_DisableIRQ(USART1_IRQn);
+  uart_ring_process(&esp_ring);
+  NVIC_EnableIRQ(USART1_IRQn);
+}
+
+void USART2_IRQHandler(void) {
+  NVIC_DisableIRQ(USART2_IRQn);
+  uart_ring_process(&debug_ring);
+  NVIC_EnableIRQ(USART2_IRQn);
+}
+
+void USART3_IRQHandler(void) {
+  NVIC_DisableIRQ(USART3_IRQn);
+  uart_ring_process(&lin2_ring);
+  NVIC_EnableIRQ(USART3_IRQn);
+}
+
+void UART5_IRQHandler(void) {
+  NVIC_DisableIRQ(UART5_IRQn);
+  uart_ring_process(&lin1_ring);
+  NVIC_EnableIRQ(UART5_IRQn);
+}
