+++ board/uart.h
@@ -1,70 +1,48 @@
-#define __DIV(_PCLK_, _BAUD_)                        (((_PCLK_)*25)/(4*(_BAUD_)))
-#define __DIVMANT(_PCLK_, _BAUD_)                    (__DIV((_PCLK_), (_BAUD_))/100)
-#define __DIVFRAQ(_PCLK_, _BAUD_)                    (((__DIV((_PCLK_), (_BAUD_)) - (__DIVMANT((_PCLK_), (_BAUD_)) * 100)) * 16 + 50) / 100)
-#define __USART_BRR(_PCLK_, _BAUD_)              ((__DIVMANT((_PCLK_), (_BAUD_)) << 4)|(__DIVFRAQ((_PCLK_), (_BAUD_)) & 0x0F))
+#ifndef PANDA_UART_H
+#define PANDA_UART_H
 
-void uart_set_baud(USART_TypeDef *u, int baud) {
-  if (u == USART1) {
-    // USART1 is on APB2
-    u->BRR = __USART_BRR(48000000, baud);
-  } else {
-    u->BRR = __USART_BRR(24000000, baud);
-  }
-}
+#define __DIV(_PCLK_, _BAUD_)       (((_PCLK_)*25)/(4*(_BAUD_)))
+#define __DIVMANT(_PCLK_, _BAUD_)   (__DIV((_PCLK_), (_BAUD_))/100)
+#define __DIVFRAQ(_PCLK_, _BAUD_)   (((__DIV((_PCLK_), (_BAUD_)) - (__DIVMANT((_PCLK_), (_BAUD_)) * 100)) * 16 + 50) / 100)
+#define __USART_BRR(_PCLK_, _BAUD_) ((__DIVMANT((_PCLK_), (_BAUD_)) << 4)|(__DIVFRAQ((_PCLK_), (_BAUD_)) & 0x0F))
 
-void uart_init(USART_TypeDef *u, int baud) {
-  // enable uart and tx+rx mode
-  u->CR1 = USART_CR1_UE;
-  uart_set_baud(u, baud);
-  
-  u->CR1 |= USART_CR1_TE | USART_CR1_RE;
-  //u->CR2 = USART_CR2_STOP_0 | USART_CR2_STOP_1;
-  //u->CR2 = USART_CR2_STOP_0;
-  // ** UART is ready to work **
+typedef struct uart_ring {
+  uint8_t w_ptr_tx;
+  uint8_t r_ptr_tx;
+  uint8_t elems_tx[FIFO_SIZE];
+  uint8_t w_ptr_rx;
+  uint8_t r_ptr_rx;
+  uint8_t elems_rx[FIFO_SIZE];
+  USART_TypeDef *uart;
+  void (*callback)(struct uart_ring*);
+} uart_ring;
 
-  // enable interrupts
-  u->CR1 |= USART_CR1_RXNEIE;
-}
+extern int has_external_debug_serial;
 
-void putch(const char a) {
-  if (has_external_debug_serial) {
-    putc(&debug_ring, a);
-  } else {
-    injectc(&debug_ring, a);
-  }
-}
+extern uart_ring debug_ring;
 
-int puts(const char *a) {
-  for (;*a;a++) {
-    if (*a == '\n') putch('\r');
-    putch(*a);
-  }
-  return 0;
-}
+uart_ring *get_ring_by_number(int a);
 
-void puth(unsigned int i) {
-  int pos;
-  char c[] = "0123456789abcdef";
-  for (pos = 28; pos != -4; pos -= 4) {
-    putch(c[(i >> pos) & 0xF]);
-  }
-}
+void uart_set_baud(USART_TypeDef *u, int baud);
 
-void puth2(unsigned int i) {
-  int pos;
-  char c[] = "0123456789abcdef";
-  for (pos = 4; pos != -4; pos -= 4) {
-    putch(c[(i >> pos) & 0xF]);
-  }
-}
+void uart_init(USART_TypeDef *u, int baud);
 
-void hexdump(void *a, int l) {
-  int i;
-  for (i=0;i<l;i++) {
-    if (i != 0 && (i&0xf) == 0) puts("\n");
-    puth2(((unsigned char*)a)[i]);
-    puts(" ");
-  }
-  puts("\n");
-}
+void putch(const char a);
 
+int puts(const char *a);
+
+void puth(unsigned int i);
+
+void puth2(unsigned int i);
+
+void hexdump(void *a, int l);
+
+int getc(uart_ring *q, char *elem);
+
+int putc(uart_ring *q, char elem);
+
+int injectc(uart_ring *q, char elem);
+
+void uart_ring_process(uart_ring *q);
+
+#endif
